+++ Changelog.md
@@ -6,6 +6,7 @@
 
   - Fixing window finder cursor display when DPI != 100%
   - [#203](../../issues/203) - The calling thread cannot access this object because a different thread owns it.
+  - [#207](../../issues/207) - Exception when trying to snoop application with invalid resource definitions inside ResourceDictionary
   
 ## 3.0.0